requirements.txt
aioredis==2.0.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.54
botocore==1.29.54
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==4.13.0
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.6.0
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.24.1
outcome==1.2.0
pandas==1.5.3
protobuf==3.20.3
pydantic==1.10.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
redis==4.4.2
requests==2.28.2
s3transfer==0.6.0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.22.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wsproto==1.2.0
zipp==3.12.1
